TODO-React Day 7
Added contracts to README
Cleaned up tests to reflect contract

We decided since this is TDD corner that we don't want to embrace tests
that don't cause us to change the code under test.  This led to the idea
that we should have a common contracts document and we updated it with
the decisions we've made so far.

Co-authored-by: Jason Grosz <grosz@testdouble.com>
Co-authored-by: Landon Gray <landon.gray@testdouble.com>