TODO-React Day 8
We started out on th requirement that the check-all checkbox check all
the todo items.  We quickly ran into the problem that our todo items are
just text and thus don't have checkboxes.

We're in the middle of a refactor to move from "naked" <li>s in the Todo
component to a new TodoItem component.  Checkbox is in place, next up is
adding the text and then eventually plugging it into <Todo>

Co-authored-by: Jason Grosz <grosz@testdouble.com>
Co-authored-by: Landon Gray <landon.gray@testdouble.com>
Co-authored-by: Andrew Walker <andrew.walker@gmail.com>