TODO-React Day 11
Completed todo list component refactor to take text and checked value.
Went back to check-all and decided to make the trigger behavior a
callback.
This put us back up at app and now we need to figure out how to assert
on the state after the checkbox is checked.

One idea is to start the app from the beginning, add all the items
"manually" then check the box and verify via the DOM everything is now
checked.

An alternate idea was to expose the "check all handler" as a pure
function that just manipulates an array.  Since we know the todo list
renders correctly if the passed in props are correct we just need to
verify that we can create the correct set of props.

So our next obstacle is decide where to put a pure function.

Co-authored-by: Andrew Walker <andrew.walker@gmail.com>
Co-authored-by: Dale Karp <dale.karp@testdouble.com>
Co-authored-by: Ed Toro <ed@testdouble.com>
Co-authored-by: Camilo Payan <camilopayan@users.noreply.github.com>
Co-authored-by: Landon Gray <landon.gray@testdouble.com>