TODO-React Day 12
After a lot of discussion we landed on the idea of a useTodos module
that will eventually grow into a hook.  For now it will be pure
functions that we can invoke.  We re-built addTodoItem with this
direction in mind.

Next up is to figure out what's next up.  We're still scared of
check-all, so we'll have to figure out if theres another state
manipulation that makes sense.  We also have some behavior around the
input box that might be worth investigating.

Co-authored-by: Andrew Walker <andrew.walker@gmail.com>
Co-authored-by: Ed Toro <ed@testdouble.com>
Co-authored-by: Landon Gray <landon.gray@testdouble.com>