Try to update react-testing-library
We're not sure if our failing test is caused by the user-event library
and how it interacts with the components.  We found a solution using
`userEvent.setup` but our libraries are too old for that.  When we
updated just `user-event` things went poorly (getByRole is broken)?

We wanted to try updating all the RTL packages and see where that goes.

Co-authored-by: Ed Toro <ed@testdouble.com>
Co-authored-by: Andrew Walker <andrew.walker@testdouble.com>
Co-authored-by: Dale Karp <dale.karp@testdouble.com>
Co-authored-by: Camilo Payan <camilopayan@users.noreply.github.com>