tdd-me-plz Day 20
More login fun.  Decided that `session#create` was all wrong, and that
we needed a `start_login` method instead.  This is where we'll kick off
the call to head to google.

All tests passing.

Next test idea is to work on "registering" aka what happens when google
calls us back with a payload?  The idea is to write the test that will
force us to plug in omniauth.

Co-authored-by: Ed Toro <ed@testdouble.com>
Co-authored-by: Eve Ragins <emragins@gmail.com>
Co-authored-by: Andrew Walker <andrew.walker@testdouble.com