Telephone Keypad Day 9
Continue to struggle with setTimeout and testing. Our theory was because
each test was invoking setTimeout (by clicking) but not clearing when
the test ended, we were seeing test polluting behavior.  So what if we
cleared the timeout at the end of the test - oh yeah that's what jest
does with its mock timer stuff.

This runner seems to use mocha and sinon so we introduced (made
visible?) sinon faketimers but don't think it's actually doing anything
useful.

Next time: Decide if we want to keep doing sinon or try something else
or roll our own setTimeout clearer or what.

Co-authored-by: Ed Toro <ed@testdouble.com>
Co-authored-by: Jason Grosz <jason.grosz@testdouble.com>